requirement.txt
Flask
google-generativeai
pandas
plotly
